requirements.txt
flask
tensorflow
tldextract
pyzbar
pillow
scikit-learn
pandas
numpy
